requirements.txt
transformers
datasets[audio]
accelerate
torchaudio
flask